token service: Handle unrecovable, expiring tokens
Artifactory servers have a global configuration
`minimum-revocable-expiry` which dictates that any token whose expiry is
less than the global minimum cannot be revoked and must expire
naturally. [1]

If the revoke API is used on a token who expiry is less than
minimum-revocable-expiry then Artifactory appears to return a 500 status
code, although this is undocumented.

The plugin is unaware of this configuration and will optimistically
attempt to revoke tokens and assume it is no revocable if such an error
occurs.

1: https://www.jfrog.com/confluence/display/ACC/Access+Tokens#AccessTokens-GeneratingExpirableTokens